Rename module to support all androidx navigation constructs
