+++ embrace-android-instrumentation-androidx-navigation/src/main/resources/META-INF/services/io.embrace.android.embracesdk.internal.arch.InstrumentationProvider
@@ -1,0 +1,1 @@
+io.embrace.android.embracesdk.instrumentation.androidx.navigation.internal.ComposeNavigationInstrumentationProvider
