--- embrace-android-instrumentation-compose-navigation/src/main/resources/META-INF/services/io.embrace.android.embracesdk.internal.arch.InstrumentationProvider
@@ -1,1 +1,0 @@
-io.embrace.android.embracesdk.internal.instrumentation.compose.navigation.ComposeNavigationInstrumentationProvider
