+++ embrace-android-sdk/src/integrationTest/kotlin/io/embrace/android/embracesdk/fakes/ComposeNavHostActivity.kt
@@ -7,7 +7,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import io.embrace.android.embracesdk.instrumentation.navigation.rememberObservedNavController
+import io.embrace.android.embracesdk.instrumentation.androidx.navigation.rememberObservedNavController
 
 class ComposeNavHostActivity : ComponentActivity(), HasNavController {
